Remove index management from datastore clients
This removes the automatic initialization step where the
API Server created and removed indexes according to the local
configuration file in the revision's source.

It also removes unnecessary work that was done to find indexes
that were relevant to queries and mutations.
