+++ AppDB/appscale/datastore/scripts/blobstore.py
@@ -254,7 +254,7 @@ class UploadHandler(tornado.web.RequestHandler):
     """
     global datastore_path
     db = datastore_distributed.DatastoreDistributed(
-      app_id, datastore_path, require_indexes=False)
+      app_id, datastore_path)
     apiproxy_stub_map.apiproxy.RegisterStub('datastore_v3', db)
     os.environ['APPLICATION_ID'] = app_id
