+++ AppTaskQueue/appscale/taskqueue/distributed_tq.py
@@ -275,7 +275,7 @@ class DistributedTaskQueue():
     db_proxy = appscale_info.get_db_proxy()
     connection_str = '{}:{}'.format(db_proxy, str(constants.DB_SERVER_PORT))
     ds_distrib = datastore_distributed.DatastoreDistributed(
-      constants.DASHBOARD_APP_ID, connection_str, require_indexes=False)
+      constants.DASHBOARD_APP_ID, connection_str)
     apiproxy_stub_map.apiproxy.RegisterStub('datastore_v3', ds_distrib)
     os.environ['APPLICATION_ID'] = constants.DASHBOARD_APP_ID
